[BE] refactor: Config 패키지 분리, 스타일 적용 (#957)
* chore: use `@Slf4j` instead of `LoggerFactory.getLogger`

* refactor: `WebConfig`가 하나의 역할만 하도록 수정

* style: dto 불필요한 개행 제거

* refactor: Config는 상위 패키지, Resolver/Interceptor는 각 도메인에 두기
