+++ backend/src/main/java/reviewme/config/cors/CorsConfig.java
@@ -1,17 +1,15 @@
-package reviewme.config;
+package reviewme.config.cors;
 
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+@Slf4j
 public class CorsConfig {
-
-    private static final Logger log = LoggerFactory.getLogger(CorsConfig.class);
 
     private CorsConfig() {
     }
