+++ backend/src/main/java/reviewme/global/exception/FieldErrorResponse.java
@@ -1,8 +1,4 @@
 package reviewme.global.exception;
 
-public record FieldErrorResponse(
-        String field,
-        Object value,
-        String message
-) {
+public record FieldErrorResponse(String field, Object value, String message) {
 }
