+++ backend/src/main/java/reviewme/review/service/ReviewRegisterService.java
@@ -1,8 +1,6 @@
 package reviewme.review.service;
 
 import lombok.RequiredArgsConstructor;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import reviewme.review.domain.Review;
@@ -14,8 +12,6 @@ import reviewme.review.service.validator.ReviewValidator;
 @Service
 @RequiredArgsConstructor
 public class ReviewRegisterService {
-
-    private static final Logger log = LoggerFactory.getLogger(ReviewRegisterService.class);
 
     private final ReviewMapper reviewMapper;
     private final ReviewValidator reviewValidator;
